src/docker/pyapi/src/requirements.txt
Flask == 1.1.2
tensorflow == 2.3.1
psycopg2-binary == 2.8.6